Moved services to their own name space and created FileUtils to save files from messages
