+++ keeganstudios.possebot/CommandModules/Help.cs
@@ -1,5 +1,6 @@
 ﻿using Discord;
 using Discord.Commands;
+using keeganstudios.possebot.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
