+++ keeganstudios.possebot/CommandModules/Theme.cs
@@ -1,11 +1,11 @@
 ﻿using Discord;
 using Discord.Commands;
 using keeganstudios.possebot.Models;
+using keeganstudios.possebot.Services;
 using keeganstudios.possebot.Utils;
 using System;
 using System.IO;
 using System.Linq;
-using System.Net.Http;
 using System.Threading.Tasks;
 
 namespace keeganstudios.possebot.CommandModules
@@ -13,18 +13,18 @@ namespace keeganstudios.possebot.CommandModules
     public class Theme : ModuleBase<SocketCommandContext>
     {        
         private readonly IAudioService _audioService;
-        private readonly IOptionsService _optionsService;
-        private readonly HttpClient _httpClient;
+        private readonly IOptionsService _optionsService;        
         private readonly ICommandUtils _commandUtils;
+        private readonly IFileUtils _fileUtils;
 
         private string[] _acceptedAudioFileExtensions = { ".mp3" };
 
-        public Theme(IAudioService audioService, IOptionsService optionsService, HttpClient httpClient, ICommandUtils commandUtils)
+        public Theme(IAudioService audioService, IOptionsService optionsService, ICommandUtils commandUtils, IFileUtils fileUtils)
         {     
             _audioService = audioService;
-            _optionsService = optionsService;
-            _httpClient = httpClient;
+            _optionsService = optionsService;            
             _commandUtils = commandUtils;
+            _fileUtils = fileUtils;
         }       
 
         [Command("ping")]
@@ -130,23 +130,8 @@ namespace keeganstudios.possebot.CommandModules
                     return;
                 }
 
-                var filePath = Path.Combine(Directory.GetCurrentDirectory(), "files", attachment.Filename);
-
-                if (!Directory.Exists(Path.GetDirectoryName(filePath)))
-                {
-                    Directory.CreateDirectory(Path.GetDirectoryName(filePath));
-                }
-
-                if (File.Exists(filePath))
-                {
-                    File.Delete(filePath);
-                }
-
-                using (var file = await _httpClient.GetStreamAsync(attachment.Url))
-                using (var fileStream = new FileStream(filePath, FileMode.CreateNew, FileAccess.Write))
-                {
-                    await file.CopyToAsync(fileStream);
-                }
+                var filePath = Path.Combine(_fileUtils.BuildAudioFilePath(Context.Guild.Id) , attachment.Filename);
+                await _fileUtils.SaveAudioFile(filePath, attachment.Url);                
 
                 var theme = new ThemeDetails
                 {
