+++ keeganstudios.possebot/Utils/CommandUtils.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using keeganstudios.possebot.Services;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
